+++ lib/index.js
@@ -4,6 +4,7 @@ export { default as Manhattan } from './Manhattan';
 export { default as ManhattanFlat } from './ManhattanFlat';
 export { default as PheWAS } from './PheWAS';
 export { default as Regional } from './Regional';
+export { default as SigSig } from './SigSig';
 export { default as GeneTrack } from './GenesTrack';
 export {
   getCytoband,
